Add Type(), IsNil(), & Blob() to type Value
Previously it was impossible to access the sqlite3_value_type() function
for a Value. Now this is exposed via Value.Type().

Additionally, it was impossible to know if a returned Value was actually
nil, which can occur when ChangesetIter.New() has no change for the
given column. Using any of the returned Value's methods would result in
panic. Now this can be checked using Value.IsNil().

Also, Value.Blob() is now implemented.

Issue #61